Build equivilant in Ruby
Didn't get a responce from Chetan. Chris didn't know. So, I built the
equivilant in Ruby.

Makes use of Wordpress Core's JSON API to grab posts
tags with `event`.

The content of this page is only rendered at build time. The benefit:
much faster load times for this page. The drawback: build must be run
each time we get a new event.

Ultimately, it would be good to cron the build.

Fixes #74